Working with authentication and permissions
Up to this moment a logged user can add, update or even delete a snippet
that was not created by himself. It will change after "Object level
permissions" is implemented.

This commit is valid to compare the changes that were made in order to
make the permissions control more granular.